requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pandas
numpy
pdfplumber
sentence-transformers
faiss-cpu
langchain
langgraph
requests